# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/AppInfo/App_detail": {
            "get": {
                "description": "Get service detail.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AppInfo Management"
                ],
                "summary": "Get service detail",
                "operationId": "/app/app_detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "service id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/db.AppInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/service/service_add_http": {
            "post": {
                "description": "Add HTTP service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service Management"
                ],
                "summary": "Add HTTP service",
                "operationId": "/service/service_add_http",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/io.ServiceAddHTTPInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/service/service_delete": {
            "get": {
                "description": "Delete a service.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service Management"
                ],
                "summary": "Delete a service",
                "operationId": "/service/service_delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "service id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/service/service_list": {
            "get": {
                "description": "Show service lists.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service Management"
                ],
                "summary": "Service list",
                "operationId": "/service/service_list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Searching keyword",
                        "name": "info",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Entries per page",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page No.",
                        "name": "page_no",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/io.ServiceListOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/service/service_stat": {
            "get": {
                "description": "Get service statistics.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service Management"
                ],
                "operationId": "/service/service_stat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "service id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/io.ServiceStatOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/service/service_update_http": {
            "post": {
                "description": "Update HTTP service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service Management"
                ],
                "summary": "Update HTTP service",
                "operationId": "/service/service_update_http",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/io.ServiceUpdateHTTPInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "db.AppInfo": {
            "type": "object",
            "properties": {
                "app_id": {
                    "type": "string"
                },
                "create_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_delete": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "qpd": {
                    "type": "integer"
                },
                "qps": {
                    "type": "integer"
                },
                "secret": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                },
                "white_ips": {
                    "type": "string"
                }
            }
        },
...
        "io.ServiceAddHTTPInput": {
            "type": "object",
            "required": [
                "ip_list",
                "rule",
                "service_desc",
                "service_name",
                "weight_list"
            ],
            "properties": {
                "black_list": {
                    "type": "string"
                },
                "client_ip_flow_limit": {
                    "type": "integer"
                },
                "head_transform": {
                    "type": "string"
                },
                "ip_list": {
                    "type": "string"
                },
                "need_https": {
                    "type": "integer"
                },
                "need_strip_uri": {
                    "type": "integer"
                },
                "need_websocket": {
                    "type": "integer"
                },
                "open_auth": {
                    "type": "integer"
                },
                "round_type": {
                    "type": "integer"
                },
                "rule": {
                    "type": "string"
                },
                "rule_type": {
                    "type": "integer"
                },
                "service_desc": {
                    "type": "string"
                },
                "service_flow_limit": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string"
                },
                "upstream_connect_time_out": {
                    "type": "integer"
                },
                "upstream_header_time_out": {
                    "type": "integer"
                },
                "upstream_idle_time_out": {
                    "type": "integer"
                },
                "upstream_max_idle": {
                    "type": "integer"
                },
                "url_rewrite": {
                    "type": "string"
                },
                "weight_list": {
                    "type": "string"
                },
                "white_list": {
                    "type": "string"
                }
            }
        },
...
        "io.ServiceListItemOutput": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "load_type": {
                    "type": "integer"
                },
                "qpd": {
                    "type": "integer"
                },
                "qps": {
                    "type": "integer"
                },
                "service_addr": {
                    "type": "string"
                },
                "service_desc": {
                    "type": "string"
                },
                "service_name": {
                    "type": "string"
                },
                "total_node": {
                    "type": "integer"
                }
            }
        },
        "io.ServiceListOutput": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/io.ServiceListItemOutput"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "io.ServiceStatOutput": {
            "type": "object",
            "properties": {
                "today": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "yesterday": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "io.ServiceUpdateHTTPInput": {
            "type": "object",
            "required": [
                "id",
                "ip_list",
                "rule",
                "service_desc",
                "service_name",
                "weight_list"
            ],
            "properties": {
                "black_list": {
                    "type": "string"
                },
                "client_ip_flow_limit": {
                    "type": "integer"
                },
                "head_transform": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 56
                },
                "ip_list": {
                    "type": "string",
                    "example": "127.0.0.1:80"
                },
                "need_https": {
                    "type": "integer"
                },
                "need_strip_uri": {
                    "type": "integer"
                },
                "need_websocket": {
                    "type": "integer"
                },
                "open_auth": {
                    "type": "integer"
                },
                "round_type": {
                    "type": "integer"
                },
                "rule": {
                    "type": "string"
                },
                "rule_type": {
                    "type": "integer"
                },
                "service_desc": {
                    "type": "string",
                    "example": "test_add_http"
                },
                "service_flow_limit": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string",
                    "example": "test_add_http"
                },
                "upstream_connect_time_out": {
                    "type": "integer"
                },
                "upstream_header_time_out": {
                    "type": "integer"
                },
                "upstream_idle_time_out": {
                    "type": "integer"
                },
                "upstream_max_idle": {
                    "type": "integer"
                },
                "url_rewrite": {
                    "type": "string"
                },
                "weight_list": {
                    "type": "string",
                    "example": "50"
                },
                "white_list": {
                    "type": "string"
                }
            }
        },
...
        "middleware.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "errmsg": {
                    "type": "string"
                },
                "errno": {
                    "type": "integer"
                },
                "stack": {
                    "type": "object"
                },
                "trace_id": {
                    "type": "object"
                }
            }
        }
    }
}